Bulk download: dynamic README.txt
