+++ src/server/repository/assessmentCycle/data/index.ts
@@ -1,5 +1,6 @@
 import { deleteNodeValues } from './deleteNodeValues'
 import { getAggregatedTableData } from './getAggregatedTableData'
+import { getBulkDownload } from './getBulkDownload'
 import { getOriginalDataPointData } from './getOriginalDataPointData'
 import { getTableData } from './getTableData'
 
@@ -10,4 +11,5 @@ export const DataRepository = {
   getOriginalDataPointData,
   deleteNodeValues,
   getAggregatedTableData,
+  getBulkDownload,
 }
