--- src/server/static/dataExport/README.txt
@@ -1,14 +1,0 @@
-README
-
-The bulk download zip archive contains three comma separated files (csv). They have been named as follows:
-    1. FRA_Years_YYYY_MM_DD.csv (YYYY_MM_DD refers to the date of the download)
-    2. Intervals_YYYY_MM_DD.csv
-    2. Annual_YYYY_MM_DD.csv
-
-1. Most of the reported data are found in the file “FRA_Years*”.  Typically the data are structured in records according to the FRA reporting years: 1990, 2000, 2010, 2015 and 2020.
-2. The file “Intervals*” contains data on Forest expansion, afforestation, natural expansion, deforestation and reforestation and each data record contains information for one interval 1990-2000, 2000-2010, 2010-2015 or 2015-2020.
-3. The file “Annual*” contains data on forest disturbances and the data are structured as annual records for the period 2000-2017.
-
-Required citation: FAO. 2020. Global Forest Resources Assessment 2020
-
-Contact: fra@fao.org
